3D Visualization and Simulation for the TERMES Project
Our lab cooperate with Harvard University. The goal in the TERMES
project is to develop a swarm construction system in which robots
cooperate with each other to build 3D structures that are much larger
than themselves. Using C++ and OpenGL to simulate the building process
and movement of robots of adding and removing blocks.